Last changes (#45)
# *Name of the PR*

## :recycle: Current situation & Problem
Okay, we went have a solid working demo. And it now is chatting with the
real db.
Can all create an account or use sample
tak@g.com
password

Mena: There is an error with the input form pushing to the db. It only
pushes the last set. so if you say last set on set 1 it pushes that or
the same for set 2, but if you go to set 3 it automatically exits
without updating the db.
Tulika: I said that I ate eggs. Then I went to my home screen and it
showed that value changed on the protein, but if I go back and forth
between home and a different page it will keep incrementing by the
previous protein amount.

When it clicks on go to LLM it should not do a linknavigation but should
instead change the environmental variable tabs. You can take a look at
how I did it on exercise week for ideas.

Yanav: You need to update the images on the workoutButton. 
You also need to make it so the video is being dynamically chosen. I
made it so that there is a video attribute to each menuItem value, but
it isn't being used.


## :gear: Release Notes 
*Add a bullet point list summary of the feature and possible migration
guides if this is a breaking change so this section can be added to the
release notes.*
*Include code snippets that provide examples of the feature implemented
or links to the documentation if it appends or changes the public
interface.*


## :books: Documentation
*Please ensure that you properly document any additions in conformance
to [Spezi Documentation
Guide](https://github.com/StanfordSpezi/.github/blob/main/DOCUMENTATIONGUIDE.md).*
*You can use this section to describe your solution, but we encourage
contributors to document your reasoning and changes using in-line
documentation.*


## :white_check_mark: Testing
*Please ensure that the PR meets the testing requirements set by CodeCov
and that new functionality is appropriately tested.*
*This section describes important information about the tests and why
some elements might not be testable.*


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md):
- [ ] I agree to follow the [Code of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md).

---------

Co-authored-by: Mena Hassan <34526502+menahassan@users.noreply.github.com>
Co-authored-by: Vishnu Ravi <vishnur@stanford.edu>